Correct screen shows depending on bluetooth-- option
Used a file "widgets.kv" to store all of the needed widgets, then two separate main files called "mainBT.kv" and "mainNoBT.kv", the only difference between the two is that one loads the normal login screen (mainNoBT), while the other loads the one used to login using bluetooth. Each one imports the "widgets.kv" file and loads depending on the "bluetooth" option in the config file.
I have done it this way because of the way the screen manager is initialized, and how I am handling the key between screens (using a StringObject).