Fixed BT server, so it only has to run once. Vnice
Before I was shutting down the server when the login was done, as it caused a segmentation fault. I found out today that you can just use the screen manager to get each screen, so i executed the BT server in MainScreen from the LoginScreenBT. This means that when the screen switches, the thread is running in MainScreen, but is being used for login.