Fixed AES and caught up on Write-up
Had to update technical solution. Now on testing again... AES had a problem where the expandedKey would get encrypted when encrypting a file, as the key is supposed to get encrypted and appended to the start of the file, however the 'originalKey' variable I made was actually a reference to the start of expanded keys, so the first 16 bytes of the expanded key would get encrypted rather than what was supposed to be a copy of the key.